+++ Assembler/Spect.Net.Assembler/Assembler/BinarySegment.cs
@@ -23,6 +23,16 @@ namespace Spect.Net.Assembler.Assembler
         public bool IsDisplaced => Displacement.HasValue;
 
         /// <summary>
+        /// HEX Displacement of the this segment
+        /// </summary>
+        public ushort? HDisplacement { get; set; }
+
+        /// <summary>
+        /// Indicates if this segment is HEX displaced
+        /// </summary>
+        public bool IsHDisplaced => HDisplacement.HasValue;
+
+        /// <summary>
         /// Emitted Z80 binary code
         /// </summary>
         public List<byte> EmittedCode { get; set; } = new List<byte>(1024);