+++ Assembler/Spect.Net.Assembler/Z80AsmVisitor.cs
@@ -275,6 +275,25 @@ namespace Spect.Net.Assembler
         }
 
         /// <summary>
+        /// Visit a parse tree produced by <see cref="Z80AsmParser.hdispPragma"/>.
+        /// <para>
+        /// The default implementation returns the result of calling <see cref="AbstractParseTreeVisitor{Result}.VisitChildren(IRuleNode)"/>
+        /// on <paramref name="context"/>.
+        /// </para>
+        /// </summary>
+        /// <param name="context">The parse tree.</param>
+        /// <return>The visitor result.</return>
+        public override object VisitHdispPragma(Z80AsmParser.HdispPragmaContext context)
+        {
+            if (IsInvalidContext(context)) return null;
+
+            return AddLine(new HDispPragma
+            {
+                Expr = (ExpressionNode)VisitExpr(context.GetChild(1) as Z80AsmParser.ExprContext)
+            }, context);
+        }
+
+        /// <summary>
         /// Visit a parse tree produced by <see cref="Generated.Z80AsmParser.equPragma"/>.
         /// </summary>
         /// <param name="context">The parse tree.</param>