+++ VsIntegration/Spect.Net.VsPackage/Z80Programs/Export/ExportFormat.cs
@@ -6,6 +6,7 @@
     public enum ExportFormat
     {
         Tap,
-        Tzx
+        Tzx,
+        IntelHex
     }
 }